added textbox and UI elements
Doors are now drawn as orange squares
Player's letters are centered in their boxes better
A text box is at the bottom of the screen, can show how many moves the user has there. Need to change inputString and inputNum to wait for the user to press a button to move (or click). 
To output to the textBox on the UI use getTextOutputArea().setText(string);